refactor: rename pastas components
