+++ browser-extension/app/components/pastas/list/pastas-list.vue
@@ -5,7 +5,7 @@
     class="flex flex-col divide-y overflow-y-auto border bg-base-100 text-base-content"
     @contextmenu="onContextMenu"
   >
-    <pastbl-pasta
+    <list-pasta
       v-for="(pasta, index) in pastas"
       :key="pasta.id"
       class="border-r"
@@ -19,7 +19,7 @@ import type { XPasta } from "~/utils/pastas.store";
 import type { Nullish } from "~/utils/types";
 import { findPasta, getCoords } from "~/utils/pastbl-pastas-list";
 import { useInfiniteRemotePastas } from "~/composables/useInfiniteRemotePastas";
-import PastblPasta from "~/components/pastas/pastbl-pasta.vue";
+import ListPasta from "~/components/pastas/list/list-pasta.vue";
 
 const props = defineProps<{
   pastas: XPasta[];
