+++ browser-extension/app/components/pastas/pasta-actions-hint.vue
